dont kill console app in integration test if it's already dead

+++ ZooIntegrationTest/ZooConsoleRunner.cs
@@ -29,7 +29,10 @@ namespace ZooIntegrationTest
 
         public void KillZooConsoleApp()
         {
-            ZooConsoleApp.Kill();
+            if (!ZooConsoleApp.HasExited)
+            {
+                ZooConsoleApp.Kill();
+            }
         }
 
         public async Task<string> ReadLineAsync()
